Added Timer class, added respawn and shield animations

+++ src/Game/Level.cpp
@@ -138,7 +138,7 @@ void Level::render() const
 		}
 	}
 }
-void Level::update(const uint64_t delta)
+void Level::update(const double delta)
 {
 	for (const auto& currentMapObject : m_levelObjects)
 	{
