+++ src/Renderer/Sprite.h
@@ -21,14 +21,14 @@ namespace RenderEngine
 	public:
 
 		struct FrameDescription {
-			FrameDescription(const glm::vec2 _leftBottomUV, glm::vec2 _rightTopUV, uint64_t _duration)
+			FrameDescription(const glm::vec2 _leftBottomUV, glm::vec2 _rightTopUV, double _duration)
 				: leftBottomUV(_leftBottomUV)
 				, rightTopUV(_rightTopUV)
 				, duration(_duration)
 			{}
 			glm::vec2 leftBottomUV;
 			glm::vec2 rightTopUV;
-			uint64_t duration;
+			double duration;
 		};
 
 		Sprite(std::shared_ptr<Texture2D> pTexture,
@@ -43,7 +43,7 @@ namespace RenderEngine
 
 
 		void insertFrames(std::vector<FrameDescription> framesDescriptions);
-		uint64_t getFrameDuration(const size_t frameId) const;
+		double getFrameDuration(const size_t frameId) const;
 		size_t getFramesCount() const;
 
 	protected:
